moved test files into _tests_ folder within src folder. updated readme
