+++ README.md
@@ -4,23 +4,36 @@ Group Members: Kylar Posey / Kyle Rivera / Paul Negrido
 
 Video Link:
 Stage1 - https://youtu.be/-7ZLHhF5aPQ
+Stage2 - TBD
 
-Local host run:
-To run the code locally do the following: 
-- Go the stock-dashboard directory
+Local host run - To run the code locally do the following: 
+Front End:
+- Go to the stock-dashboard directory.
 - Install the required modules with npm install
-- run the script "npm start" 
+- run the script "npm start"
 
-To run the hosted app in firebase, follow the link: 
+Back End:
+- Go to the backend directory.
+- Install the required modules with npm install
+- Start the server with the command: node index.js
+
+Or, go to the app hosted firebase with the link below: 
 https://group5-capstone-project.web.app/
 
 Running tests:
+Jest Testing:
 To run tests locally do the following:
 - Go the stock-dashboard directory.
-- To run all jest unit tests, run the command with "all" flag. "
-   Command ==> npm test -- --all
-- To show jest unit tests coverage, run the command with the "coverage" and "all" flag.
-   Comand ==> npm test -- --coverage --all
+- To run all jest unit tests, run the command below:
+   Command ==> npm test
+- To show jest unit tests coverage, run the command below:
+   Comand ==> npm test -- --coverage
+
+TestCafe Testing (End-to-End):
+- Go the stock-dashboard directory.
+- Run the command: testcafe < installed browser of choice > ./testCafeTest/< filename of test >
+   Example: testcafe chrome ./testCafeTest/testCafeInitial.js
+
 
 GitHub Actions:
 - We've setup GitHub workflows to run automated tests.
